pyproject.toml
[tool.poetry]
name = "dider213olabhubscope"
version = "0.1.0"
description = ""
authors = ["Kevin Geng <khg@andrew.cmu.edu>"]

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1.2"
sqlalchemy = "^1.3.17"
flask-sqlalchemy = "^2.4.3"
flask-migrate = "^2.5.3"
python-dotenv = "^0.13.0"
flask-login = "^0.5.0"
requests = "^2.23.0"
flask-wtf = "^0.14.3"
flask-dance = "^3.0.0"
blinker = "^1.4"
pyjwt = "^1.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
